Split apart the last few source files that were already moved into sub-packages.
