+++ core/src/main/scala/com/rackspace/com/papi/components/checker/step/base/StepContext.scala
@@ -1,0 +1,30 @@
+/***
+ *   Copyright 2014 Rackspace US, Inc.
+ *
+ *   Licensed under the Apache License, Version 2.0 (the "License");
+ *   you may not use this file except in compliance with the License.
+ *   You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ */
+package com.rackspace.com.papi.components.checker.step.base
+
+import javax.servlet.FilterChain
+
+import com.rackspace.com.papi.components.checker.handler.ResultHandler
+import com.rackspace.com.papi.components.checker.servlet._
+import com.rackspace.com.papi.components.checker.step.results.{MismatchResult, MultiFailResult, Result}
+import com.rackspace.com.papi.components.checker.util.HeaderMap
+
+import scala.collection.mutable.ListBuffer
+
+//
+//  Used to keep context about the current request
+//
+case class StepContext(uriLevel : Int = 0, requestHeaders : HeaderMap = new HeaderMap, handler: Option[ResultHandler] = None)
