+++ core/src/main/scala/com/rackspace/com/papi/components/checker/step/startend/Accept.scala
@@ -1,0 +1,53 @@
+/***
+ *   Copyright 2014 Rackspace US, Inc.
+ *
+ *   Licensed under the Apache License, Version 2.0 (the "License");
+ *   you may not use this file except in compliance with the License.
+ *   You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ */
+package com.rackspace.com.papi.components.checker.step.startend
+
+import java.util.HashMap
+import javax.servlet.FilterChain
+import javax.xml.namespace.QName
+import javax.xml.validation.Schema
+
+import com.rackspace.com.papi.components.checker.servlet._
+import com.rackspace.com.papi.components.checker.step.XSDStringValidator
+import com.rackspace.com.papi.components.checker.step.base.{ConnectedStep, Step, StepContext}
+import com.rackspace.com.papi.components.checker.step.results._
+
+import scala.util.matching.Regex
+
+//
+//  The accept state, send the request over
+//
+class Accept(id : String, label : String, val priority : Long) extends Step(id, label) {
+  override def check(req : CheckerServletRequest,
+                     resp : CheckerServletResponse,
+                     chain : FilterChain,
+                     context : StepContext) : Option[Result] = {
+    //
+    //  Send any request headers in the current context
+    //
+    req.addHeaders(context.requestHeaders)
+
+    //
+    //  For now, accept always send out to the chain
+    //
+    chain.doFilter(req, resp)
+
+    //
+    //  Send request...
+    //
+    return Some( new AcceptResult( "", context, id, priority ) )
+  }
+}
