+++ core/src/main/scala/com/rackspace/com/papi/components/checker/step/startend/Start.scala
@@ -1,0 +1,54 @@
+/***
+ *   Copyright 2014 Rackspace US, Inc.
+ *
+ *   Licensed under the Apache License, Version 2.0 (the "License");
+ *   you may not use this file except in compliance with the License.
+ *   You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ */
+package com.rackspace.com.papi.components.checker.step.startend
+
+import java.util.HashMap
+import javax.servlet.FilterChain
+import javax.xml.namespace.QName
+import javax.xml.validation.Schema
+
+import com.rackspace.com.papi.components.checker.servlet._
+import com.rackspace.com.papi.components.checker.step.XSDStringValidator
+import com.rackspace.com.papi.components.checker.step.base.{ConnectedStep, Step, StepContext}
+import com.rackspace.com.papi.components.checker.step.results._
+
+import scala.util.matching.Regex
+
+//
+//  The start step
+//
+class Start(id : String, label : String, next : Array[Step]) extends ConnectedStep(id, label, next) {
+  override def checkStep(req : CheckerServletRequest,
+                         resp : CheckerServletResponse,
+                         chain : FilterChain,
+                         context : StepContext ) : Option[StepContext] = Some(context)
+
+  override def check(req : CheckerServletRequest,
+                     resp : CheckerServletResponse,
+                     chain : FilterChain,
+                     context : StepContext) : Option[Result] = {
+    //
+    // If we have a malformed URI, then we can't even start the
+    // machine, so return a 400 result.
+    //
+    req.parsedRequestURI match {
+      case (_, Some(e)) => Some(new ErrorResult(e.getMessage(), 400, context, id, Long.MaxValue))
+      case _ => super.check(req, resp, chain, context)
+    }
+  }
+
+  override val mismatchMessage : String = "Bad Request?"
+}
